hsmodule: remove `modinit`, consolidate to `module`, annotate init seq
One of Lua's pitfalls is that everything can be a global by default, it
makes understanding the initialization process a bit of a PITA. To work
despite that, we're trying to make the best of the pseudo-OOP that Lua
provides.

We'll slowly try to streamline and document the initialization sequence,
changing it if we have to.

For instance, we're doing the settings read _before_ the macOS and Live
sanity check and _well before_ we're reading the menuconfig. The reason
will be apparent in subsequent commits.